Eliminate the `elf` module, move its contents to `image_format`
